AgentChat/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0
openai>=1.0.0
mem0ai>=0.1.0
